Fix yet more implicit conversions

+++ benchmarks/combinatoric.cpp
@@ -7,7 +7,7 @@
 
 
 static void BM_factorial(benchmark::State &state) {
-	const std::size_t n = state.range();
+	const std::size_t n = static_cast<std::size_t>(state.range());
 
 	for (auto _ : state) {
 		std::size_t result = tpack::details::factorial(n);
@@ -19,8 +19,8 @@ BENCHMARK(BM_factorial)->DenseRange(0, 20);
 
 
 static void BM_binomial(benchmark::State &state) {
-	const std::size_t n = state.range(0);
-	const std::size_t k = state.range(1);
+	const std::size_t n = static_cast<std::size_t>(state.range(0));
+	const std::size_t k = static_cast<std::size_t>(state.range(1));
 
 	for (auto _ : state) {
 		std::size_t result = tpack::details::binomial(n, k);
